move value float->int to fix floating point errors

+++ src/ai.h
@@ -13,12 +13,12 @@ public:
     static chessMove bestMove(const board &board, int depth, side side);
     static std::vector<chessMove> minimaxHeadTrace(const board &startBoard, int depth, side maxPlayerSide);
 private:
-    static std::tuple<chessMove, float> minimaxSlow(const board &startBoard, int depth, bool maxPlayer, side maxPlayerSide);
+    static std::tuple<chessMove, int> minimaxSlow(const board &startBoard, int depth, bool maxPlayer, side maxPlayerSide);
     static chessMove minimaxHead(const board &startBoard, int depth, side maxPlayerSide);
-    static float minimax(const board &startBoard, int depth, float a, float b, bool maxPlayer, side maxPlayerSide);
-    static float minimaxBase(const board &startBoard, int depth, bool maxPlayer, side maxPlayerSide);
+    static int minimax(const board &startBoard, int depth, int a, int b, bool maxPlayer, side maxPlayerSide);
+    static int minimaxBase(const board &startBoard, int depth, bool maxPlayer, side maxPlayerSide);
 
-    static std::vector<chessMove> minimaxTrace(const board &startBoard, int depth, float a, float b, bool maxPlayer, side maxPlayerSide);
+    static std::vector<chessMove> minimaxTrace(const board &startBoard, int depth, int a, int b, bool maxPlayer, side maxPlayerSide);
     static std::vector<chessMove> minimaxBaseTrace(const board &startBoard, int depth, bool maxPlayer, side maxPlayerSide);
 };
